+++ lib/src/xml/mixins/has_writer.dart
@@ -17,18 +17,30 @@ mixin XmlHasWriter implements XmlHasVisitor {
   /// If `pretty` is set to `true` the output is nicely reformatted, otherwise
   /// the tree is emitted verbatim.
   ///
-  /// The option `indent` is only used when pretty formatting to customize the
+  /// The option `indent` is used only when pretty formatting to customize the
   /// indention of nodes, by default nodes are indented with 2 spaces.
+  ///
+  /// The option `newLine` is used only when pretty formatting to customize the
+  /// printing of new lines, by default the standard new-line `'\n'` character
+  /// is used.
+  ///
+  /// The option `level` is used only when pretty formatting to customize the
+  /// initial indention level, by default this is `0`.
   String toXmlString(
       {bool pretty = false,
+      XmlEntityMapping entityMapping = const XmlDefaultEntityMapping.xml(),
+      int level = 0,
       String indent = '  ',
-      XmlEntityMapping entityMapping = const XmlDefaultEntityMapping.xml()}) {
+      String newLine = '\n'}) {
     final buffer = StringBuffer();
-    if (pretty) {
-      XmlPrettyWriter(buffer, entityMapping, 0, indent).visit(this);
-    } else {
-      XmlWriter(buffer, entityMapping).visit(this);
-    }
+    final writer = pretty
+        ? XmlPrettyWriter(buffer,
+            entityMapping: entityMapping,
+            level: level,
+            indent: indent,
+            newLine: newLine)
+        : XmlWriter(buffer, entityMapping: entityMapping);
+    writer.visit(this);
     return buffer.toString();
   }
 }